Update brunel1.py
All of the entities should have a loot list which has been assigned to them. With the exception of the commented entities, there should be no "default loot" for the entities.
Shuffled the biome spawns we spoke of yesterday, need to create the hill biomes to better distribute some of the entities, this is next.